Give the daemon-to-HNP report one writer, beside its reader
Two things, both about a message with three writers and one reader.

First, errmgr/prted's "only other state is terminated" block is
unreachable and has been since the proc states were numbered.  The
handler is registered for COMM_FAILED and ERROR, and the state machine
only falls back on the ERROR registration for states ABOVE
PRTE_PROC_STATE_ERROR (50) - every one of which is above
PRTE_PROC_STATE_TERMINATED (20) and taken by the branch before it.  A
plain TERMINATED goes to state/prted's track_procs.  What sat here was a
partial, stale copy of that: no PRTE_JOB_TERM_NOTIFIED dedup, no IOF
completion, no PMIx notification, and a bare PMIX_RELEASE(jdata).  Had
anything ever reached it, it would have been a worse version of work
already done elsewhere.  Delete it.

Second, the report itself.  PRTE_PLM_UPDATE_PROC_STATE is read in exactly
one place, prte_plm_base_recv(), and was written in three: errmgr/prted,
state/prted (byte-identical helpers, bar a TERM_REPORTED filter), and a
hand-rolled sequence in prted_abort().  The wire carries no version, so
the project rule is that a field change must touch reader and writer in
the same commit - which is only enforceable if there is one writer to
find.  Move the packers to plm/base/plm_base_receive.c, next to the
reader, with the filter as a parameter, and have both components call
them.  (prted_abort still hand-rolls its message: it describes the daemon
itself, which is not in prte_local_children, so it has no child to pack.)

The round-trip test moves with them, from test/unit/errmgr to
test/unit/plm, and gains the skip_reported case: a proc already reported
must not be sent again, and one that is sent must come back flagged.
