+++ src/mca/errmgr/base/errmgr_private.h
@@ -70,22 +70,5 @@ PRTE_EXPORT void prte_errmgr_base_log(int error_code, char *filename, int line);
  */
 PRTE_EXPORT bool prte_errmgr_base_any_live_children(const char *job);
 
-/* Pack a proc-state report for the HNP.
- *
- * These build the body of a PRTE_PLM_UPDATE_PROC_STATE message, whose reader
- * is prte_plm_base_receive().  The two must be changed together - the wire
- * carries no format version - so they are covered by a round-trip unit test
- * in test/unit/errmgr.
- *
- * prte_errmgr_base_pack_state_for_proc() packs one proc as
- * {rank, pid, state, exit_code}; prte_errmgr_base_pack_state_update() packs a
- * job's nspace, every local child of that job, and the PMIX_RANK_INVALID
- * terminator that tells the reader the job is complete.
- */
-PRTE_EXPORT int prte_errmgr_base_pack_state_for_proc(pmix_data_buffer_t *alert,
-                                                     prte_proc_t *child);
-PRTE_EXPORT int prte_errmgr_base_pack_state_update(pmix_data_buffer_t *alert,
-                                                   prte_job_t *jobdat);
-
 END_C_DECLS
 #endif
